ssddata: Species Sensitivity Distribution Data

Reference data sets of species sensitivities to compare the results of fitting species sensitivity distributions using software such as 'ssdtools' and 'Burrlioz'. It consists of 17 primary data sets from four different Australian and Canadian organizations as well as five datasets from anonymous sources. It also includes a data set of the results of fitting various distributions using different software.

See also

Author

Maintainer: Rebecca Fisher R.Fisher@aims.gov.au

Authors:

Joe Thorley joe@poissonconsulting.ca (ORCID)

Ayla Pearson ayla@poissonconsulting.ca (ORCID)

Other contributors:

Carl Schwarz [contributor]

David Fox [contributor]