Species Sensitivity Data for aluminium_marine

Species Sensitivity Data provided by the Australian Institute of Marine Science for aluminium in marine water.

aims_aluminium_marine

Format

An object of class tbl_df (inherits from tbl, data.frame) with 20 rows and 9 columns.

Details

These data were sourced from: van Dam JW, Trenfield MA, Streten C, Harford AJ, Parry D, van Dam RA (2018). “Water quality guideline values for aluminium, gallium and molybdenum in marine environments.” Environmental Science and Pollution Research, 25(26), 26592--26602. ISSN 16147499, https://link.springer.com/article/10.1007/s11356-018-2702-y.

The columns are as follows:

Common

The species common name (chr).

Conc

The chemical concentration in micrograms per Litre (dbl).

Domain

Tropical, temperate or other filter (chr).

Life_stage

Life stage of the test organism (chr).

Phylum

The Phylum name (chr).

Source

The endpoint primary data source (chr).

Species

The species names name (chr).

Test_endpoint

Endpoint statistic, EC10, NEC etc (chr).

Toxicity_measure

Type of toxicity measure used (chr).

Examples

print(aims_aluminium_marine, n=Inf)
#> # A tibble: 20 × 9
#>    Common     Conc Domain  Life_stage Phylum  Source     Species   Test_endpoint
#>    <chr>     <dbl> <chr>   <chr>      <chr>   <chr>      <chr>     <chr>        
#>  1 Diatom      610 Temper… NA         Bacill… (Gillmore… Minutoce… EC10         
#>  2 Diatom       80 Temper… NA         Bacill… (Gillmore… Ceratone… EC10         
#>  3 Diatom       18 Temper… NA         Bacill… (Gillmore… Ceratone… EC10         
#>  4 Diatom       27 Mixed   NA         Bacill… (Gillmore… Ceratone… EC10         
#>  5 Diatom     2100 Temper… NA         Bacill… (Gillmore… Phaeodac… EC10         
#>  6 Green mi…  1400 Temper… NA         Chloro… (Golding … Dunaliel… EC10         
#>  7 Green mi…  3200 Temper… NA         Chloro… (Golding … Tetrasel… EC10         
#>  8 Kelp       6800 Temper… Zoospore   Ochrop… (Golding … Ecklonia… EC10         
#>  9 Sea grape  9800 Temper… Zygote     Ochrop… (Golding … Hormosir… NOEC         
#> 10 Blue mus…   250 Temper… Embryo     Mollus… (Golding … Mytilus … EC10         
#> 11 Sydney r…   100 Temper… Embryo     Mollus… (Wilson a… Saccostr… NOEC         
#> 12 Black se… 28000 Temper… Embryo     Echino… (Golding … Heliocar… NOEC         
#> 13 Diatom       14 Tropic… NA         Bacill… (Harford … Ceratone… EC10         
#> 14 Golden m…   640 Tropic… NA         Haptop… (Trenfiel… Tisochry… EC10         
#> 15 Blacklip…   410 Tropic… Embryo     Mollus… (Golding … Saccrost… EC10         
#> 16 Dog whelk   115 Tropic… Larva      Mollus… (Trenfiel… Nassariu… EC10         
#> 17 Australi…   312 Tropic… Zoea       Crusta… (van Dam … Coenobit… EC10         
#> 18 Striped …   416 Tropic… Nauplius   Crusta… (van Dam … Amphibal… EC10         
#> 19 Branchin…  1300 Tropic… Larva      Cnidar… (Negri et… Acropora… EC10         
#> 20 Glass an…   817 Tropic… Adult      Cnidar… (Trenfiel… Exaiptas… EC10         
#> # … with 1 more variable: Toxicity_measure <chr>